rename directory to export_util
