+++ export_util/export_et.py
@@ -19,8 +19,8 @@ from executorch.backends.xnnpack.partition.xnnpack_partitioner import (
 from executorch.exir.capture._config import EdgeCompileConfig, ExecutorchBackendConfig
 from executorch.exir.passes.quant_fusion_pass import QuantFusionPass
 from executorch.exir.passes.sym_shape_eval_pass import ConstraintBasedSymShapeEvalPass
-from export.executorch_portable_utils import export_to_edge
-from export.export_et_util import replace_attention_with_custom_sdpa_attention
+from export_util.executorch_portable_utils import export_to_edge
+from export_util.export_et_util import replace_attention_with_custom_sdpa_attention
 from torch._export import capture_pre_autograd_graph
 
 default_device = "cpu"
